requirements.txt
aiohttp~=3.8.1
PyYAML~=6.0
alembic
psycopg2-binary
psycopg[binary]~=3.0.15
loguru~=0.6.0
aiohttp_jinja2